% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Filters.r
\name{apply_filters}
\alias{apply_filters}
\title{Apply a set of user-defined filters to a dataframe}
\usage{
apply_filters(df, filters)
}
\arguments{
\item{df}{The dataframe to be filtered}

\item{filters}{A set of filters in a vector, like c("col1 == TRUE", "col2 > 20")}
}
\value{
The filtered dataframe
}
\description{
Apply a set of user-defined filters to a dataframe
}
