% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Filters.r
\name{subset_df}
\alias{subset_df}
\title{Subset a dataframe to only keep rows that are also present in another dataframe}
\usage{
subset_df(df, col, df2, col2)
}
\arguments{
\item{df}{The dataframe to subset}

\item{col}{The df column to use for the subset}

\item{df2}{The dataframe to use as reference}

\item{col2}{The column to use as reference in df2}
}
\value{
The subsetted dataframe (df)
}
\description{
Subset a dataframe to only keep rows that are also present in another dataframe
}
